Add files via upload
This is the HTML file for the Pong game. The script we developed made use of Juan Cazala's Synaptic library, which needs to be downloaded at this link for the game to work https://github.com/cazala/synaptic. To download, click the green button that says "Clone or Download", then "Download ZIP", and then unzip the folder. Then, you should be able to download this HTML file to the same loaction and play it in-browser.